+++ packages/react-server/src/webpack-plugin/test/webpack-plugin.test.ts
@@ -2,7 +2,7 @@ import path from 'path';
 
 import webpack, {Compiler} from 'webpack';
 
-import {HEADER, Options} from '../webpack-plugin';
+import {HEADER, Options} from '../shared';
 
 import {withWorkspace} from './utilities/workspace';
 
@@ -225,6 +225,85 @@ describe('webpack-plugin', () => {
       BUILD_TIMEOUT,
     );
   });
+
+  describe('error component', () => {
+    it(
+      'imports the Error component in the server when error.js exists',
+      async () => {
+        const name = 'rails-includes-error-component';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+          await workspace.write(
+            `${basePath}/error.js`,
+            BASIC_JS_ERROR_COMPOENT,
+          );
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+
+    it(
+      'does not import the Error component in the server when error.js does not exists',
+      async () => {
+        const name = 'rails-doesnt-include-error-component';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).not.toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+
+    it(
+      'includes the Error component in the server when an error folder exists',
+      async () => {
+        const name = 'rails-includes-error-component-folder';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+          await workspace.write(
+            `${basePath}/error/index.js`,
+            BASIC_JS_ERROR_COMPOENT,
+          );
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+  });
 });
 
 function runBuild(configPath: string): Promise<any[]> {
@@ -288,6 +367,10 @@ const BASIC_ENTRY = `console.log('I am a bespoke entry');`;
 
 const BASIC_JS_MODULE = `module.exports = () => {
   return 'I am totally a react component';
+};`;
+
+const BASIC_JS_ERROR_COMPOENT = `module.exports = () => {
+  return 'I am totally an Error component';
 };`;
 
 const createWebpackConfig = (